Uses u32 in mock block

+++ client/statement-store/src/lib.rs
@@ -891,7 +891,7 @@ mod tests {
 	type Extrinsic = sp_runtime::OpaqueExtrinsic;
 	type Hash = sp_core::H256;
 	type Hashing = sp_runtime::traits::BlakeTwo256;
-	type BlockNumber = u64;
+	type BlockNumber = u32;
 	type Header = sp_runtime::generic::Header<BlockNumber, Hashing>;
 	type Block = sp_runtime::generic::Block<Header, Extrinsic>;
